% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbudg.R
\name{ADBUDG}
\alias{ADBUDG}
\title{Adbudg}
\usage{
ADBUDG(X, v = 1, rho = mean, p = 1, beta = 1, alpha = 0,
  begin_year = start(X)[1], begin_period = start(X)[2],
  end_year = end(X)[1], end_period = end(X)[2],
  frecuency = frequency(X))
}
\arguments{
\item{X}{Es una lista de numeros a la cual se le va a aplicar la transformacion adbook, la funcion tambien admite objetos tipo ts.}

\item{v}{Es un valor extra de la transformacion, este valor esta por default igual a 1. Si se llegase a
colocar un numero mayor a 1 en dicho parametro, la ecuacion de la nueva transformacion seria asi:
 \deqn{y = \alpha +{\beta* X^\gamma}/{X^\gamma+\rho^\gamma}}}

\item{rho}{Es la funcion que se va a establecer para calcular esta transformacion, por default es la funcion "mean".}

\item{p}{Es un valor extra que multiplica al valor de rho. El default es igual a 1.}

\item{beta}{Es un valor extra de la transformacion, este valor esta por default igual a 1}

\item{alpha}{Es un valor extra de la transformacion, este valor esta por default igual a 0}

\item{begin_year}{Anho en el cual comienza la serie (este parametro solo se activa si el objeto no es ts)}

\item{begin_period}{Periodo del Anho en el cual comienza la serie (este parametro solo se activa si el objeto no es ts)}

\item{end_year}{Anho en el cual termina la serie (este parametro solo se activa si el objeto no es ts)}

\item{end_period}{Periodo del Anho en el cual termuina la serie (este parametro solo se activa si el objeto no es ts)}

\item{frecuency}{Frecuencia del objeto (este parametro solo se activa si el objeto no es ts)}

\item{Ecuation}{\deqn{y = \alpha +{{\beta*X}/{X+ \rho}} } Se debe tener en cuenta que \eqn{\rho} puede ser el promedio, mediana, valor minimo, etc. De la variable a transformar.}
}
\value{
Un objeto ts, que es la variable X transformada en adstock
}
\description{
Adbudg
}
\examples{

a = c(1,2,3,34,5,6,7,0,0,0,2,3,4,5)
d=ADBUDG(a)






}
