% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstock.R
\name{ADSTOCK}
\alias{ADSTOCK}
\title{Adstock}
\usage{
ADSTOCK(X, rate = 0, begin_year = start(X)[1],
  begin_period = start(X)[2], end_year = end(X)[1],
  end_period = end(X)[2], frecuency = frequency(X))
}
\arguments{
\item{X}{Es un vector de numeros el cual se le va a aplicar la transformacion adstock, la funcion tambien admite objetos tipo ts.}

\item{rate}{Es la tasa con la cual se va a calcular el adstock, esta funciona con el numero entero es decir si por ejemplo se quiere calcular un adstock de 50%, se coloca 50.}

\item{begin_year}{Anho en el cual comienza la serie.}

\item{begin_period}{Periodo del Anho en el cual comienza la serie.}

\item{end_year}{Anho en el cual termina la serie.}

\item{end_period}{Periodo del Anho en el cual termina la serie.}

\item{frecuency}{Frecuencia del objeto.}
}
\value{
Un objeto ts, que es la variable X transformada en adstock.
}
\description{
Adstock
}
\examples{


a=c(1,2,3,4,5,56,6,7,4,2,3)
ADSTOCK(a,50)



}
