% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbudg.R
\name{Power}
\alias{Power}
\title{Power}
\usage{
Power(X, beta = 1, v = 1, alpha = 0, begin_year = start(X)[1],
  begin_period = start(X)[2], end_year = end(X)[1],
  end_period = end(X)[2], frecuency = frequency(X))
}
\arguments{
\item{X}{Es una lista de numeros a la cual se le va a aplicar la transformacion Power.}

\item{beta}{Es un valor extra de la transformacion, este valor esta por default igual a 1}

\item{v}{Es un valor extra de la transformacion, este valor esta por default igual a 1.}

\item{alpha}{Es un valor extra de la transformacion, este valor esta por default igual a 0}

\item{begin_year}{Anho en el cual comienza la serie (este parametro solo se activa si el objeto no es ts)}

\item{begin_period}{Periodo del Anho en el cual comienza la serie (este parametro solo se activa si el objeto no es ts)}

\item{end_year}{Anho en el cual termina la serie (este parametro solo se activa si el objeto no es ts)}

\item{end_period}{Periodo del Anho en el cual termuina la serie (este parametro solo se activa si el objeto no es ts)}

\item{frecuency}{Frecuencia del objeto (este parametro solo se activa si el objeto no es ts)}

\item{Ecuation}{\deqn{y = \alpha +{\beta*(X^\gamma)} }}
}
\value{
la variable transformada, un vector tipo ts.
}
\description{
Power
}
\examples{

a = c(1,2,3,34,5,6,7,8,10,1000,2,3,4,5)
Power(a)



}
